# server/web/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tiny Log API",
    "version": "1.0.0",
    "description": "Test log ingest and host/service monitoring.\n\n**Write APIs** need `Authorize` with Bearer API key (`TINY_LOG_API_KEY`) or client token (`TINY_LOG_CLIENT_TOKEN`).\n\n**Read APIs** need an admin session cookie (login first) unless `TINY_LOG_AUTH_MODE=anonymous`."
  },
  "servers": [
    { "url": "/", "description": "Current host" }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_KEY",
        "description": "Server API key for POST /api/v1/logs* and POST /api/v1/metrics/batch"
      },
      "ClientToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "CLIENT_TOKEN",
        "description": "Browser client token for POST /api/v1/client/logs only"
      },
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tiny_log_session",
        "description": "Set automatically after POST /api/auth/login (use Authorize or login via UI)"
      }
    },
    "schemas": {
      "NewLog": {
        "type": "object",
        "required": ["app", "level", "message"],
        "properties": {
          "app": { "type": "string", "example": "wordyguru" },
          "level": {
            "type": "string",
            "enum": ["debug", "info", "warn", "error", "fatal"],
            "example": "error"
          },
          "message": { "type": "string", "example": "MySQL timeout" },
          "timestamp": {
            "type": "string",
            "example": "2026-09-16T15:20:31.123Z",
            "description": "RFC3339, unix ms, or unix seconds (numbers below 1e11 are seconds; fractions ok). Omitted = server UTC now. Before 1970 or more than 24h ahead = server UTC now"
          },
          "source": { "type": "string", "example": "server" },
          "meta": {
            "type": "object",
            "additionalProperties": true,
            "example": { "route": "/search", "status": 504 }
          }
        }
      },
      "BatchLogs": {
        "type": "object",
        "required": ["logs"],
        "properties": {
          "logs": {
            "type": "array",
            "maxItems": 500,
            "items": { "$ref": "#/components/schemas/NewLog" }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
      "PayloadTooLarge": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "payload too large" },
          "max_batch": {
            "type": "integer",
            "example": 500,
            "description": "Most entries per request (services, for /metrics/batch)"
          },
          "max_body_bytes": {
            "type": "integer",
            "example": 1048576,
            "description": "Request body cap (TINY_LOG_MAX_BODY_MB)"
          }
        }
      },
      "BatchRejected": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "invalid level: loud" },
          "rejected": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "index": { "type": "integer", "example": 1 },
                "error": { "type": "string", "example": "invalid level: loud" }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/auth/setup": {
      "post": {
        "tags": ["auth"],
        "summary": "First-run: create the admin (needs the one-time setup token)",
        "description": "Only while no admin exists. The token is TINY_LOG_SETUP_TOKEN, or generated at start and printed in the server log as /setup#token=…. It is spent on success. Until setup completes, every session-protected endpoint returns 401 in login mode.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["setup_token", "username", "password", "confirm_password"],
                "properties": {
                  "setup_token": { "type": "string" },
                  "username": { "type": "string", "maxLength": 64 },
                  "password": { "type": "string", "minLength": 12 },
                  "confirm_password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Admin created; session cookie set" },
          "400": { "description": "Invalid username/password" },
          "401": { "description": "Missing or wrong setup token" },
          "409": { "description": "Admin already exists" },
          "429": { "description": "Too many wrong tokens from this IP (5 per 5 min)" }
        }
      }
    },
...
    "/api/v1/logs": {
      "post": {
        "tags": ["logs"],
        "summary": "Write a single log",
        "security": [{ "ApiKey": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NewLog" }
            }
          }
        },
        "responses": {
          "200": { "description": "Inserted" },
          "400": {
            "description": "Malformed JSON or invalid entry",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "413": { "description": "Too large — split and resend; body gives the limits", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PayloadTooLarge" } } } },
          "415": { "description": "Content-Type is not application/json", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "422": { "description": "JSON does not match the schema (e.g. missing field)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "get": {
        "tags": ["logs"],
        "summary": "List / search logs",
        "security": [{ "SessionCookie": [] }],
        "parameters": [
          { "name": "app", "in": "query", "schema": { "type": "string" } },
          {
            "name": "level",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["debug", "info", "warn", "error", "fatal"]
            }
          },
          { "name": "source", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "from", "in": "query", "schema": { "type": "string" } },
          { "name": "to", "in": "query", "schema": { "type": "string" } },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 100, "maximum": 500 }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": { "type": "integer", "default": 0 }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/v1/logs/batch": {
      "post": {
        "tags": ["logs"],
        "summary": "Write a batch of logs",
        "description": "All or nothing: one invalid entry rejects the whole batch with 400 and `rejected` lists every bad index — drop those and resend the rest. 413 = too many entries or bytes; its body gives `max_batch` and `max_body_bytes`. Every error body is JSON.",
        "security": [{ "ApiKey": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BatchLogs" },
              "example": {
                "logs": [
                  {
                    "app": "wordyguru",
                    "level": "info",
                    "message": "request completed"
                  },
                  {
                    "app": "nginx",
                    "level": "error",
                    "message": "upstream timeout"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "All inserted" },
          "400": {
            "description": "Malformed JSON, or invalid entries (nothing stored)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BatchRejected" } } }
          },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "413": { "description": "Too large — split and resend; body gives the limits", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PayloadTooLarge" } } } },
          "415": { "description": "Content-Type is not application/json", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "422": { "description": "JSON does not match the schema (e.g. missing field)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/v1/logs/{id}": {
      "get": {
        "tags": ["logs"],
        "summary": "Get one log by id",
        "security": [{ "SessionCookie": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Not found" }
        }
      }
    },
...
  }
}